+++ lwk_wollet/src/pset_create.rs
@@ -1,5 +1,4 @@
 use crate::bitcoin::PublicKey as BitcoinPublicKey;
-use crate::contract::Contract;
 use crate::elements::confidential::AssetBlindingFactor;
 use crate::elements::issuance::ContractHash;
 use crate::elements::pset::{Output, PartiallySignedTransaction};
@@ -9,6 +8,7 @@ use crate::hashes::Hash;
 use crate::model::{Recipient, WalletTxOut};
 use crate::tx_builder::add_input_inner;
 use crate::wollet::Wollet;
+use crate::Contract;
 use crate::Network;
 use elements::pset::elip100::{AssetMetadata, TokenMetadata};
 use std::collections::HashMap;